split password/token-related security functions into security.pwdz and security.tokenz
